JIRA: RW-845 Make r623392 (std::stable_sort -> list.sort) permanent.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@92369 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/gtf_location_merger.hpp
@@ -76,9 +76,10 @@ public:
     TSeqPos mStop;
     ENa_strand mStrand;
     string mType;
-    size_t mPartNum; 
+    int mPartNum; 
 
 };
+
 
 //  ============================================================================
 class CGtfLocationMerger
